project restructuring, add sass
