+++ Valentina_Aktanko/src/components/Message/Message.jsx
@@ -1,3 +1,5 @@
+import './Message.scss';
+
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
