+++ Valentina_Aktanko/src/components/Message/Message.scss
@@ -1,0 +1,4 @@
+.message {
+    background-color: aquamarine;
+    width: percentage(1/10);
+}
