+++ Valentina_Aktanko/src/components/Message/index.js
@@ -1,0 +1,1 @@
+export * from './Message';
