+++ Valentina_Aktanko/src/components/MessagesList/index.js
@@ -1,0 +1,1 @@
+export * from './MessagesList';
