+++ Valentina_Aktanko/src/components/Messenger/index.js
@@ -1,0 +1,1 @@
+export * from './Messenger';
